fix: move auth.ts, middleware.ts to correct location
