+++ src/app/layout.tsx
@@ -1,7 +1,7 @@
+import { auth } from '@/auth';
 import type { Metadata } from 'next';
 import { SessionProvider } from 'next-auth/react';
 import { Inter } from 'next/font/google';
-import { auth } from '../../auth';
 import './globals.css';
 
 const inter = Inter({ subsets: ['latin'] });
